Log extra stuff. Also move dynamo type to common types, and use it in the callback lambda
